move the remains of grizzly.users.base to grizzly.users
remove SftpUser, never used and does not work as intended as it was.
remove unused interfaces, HttpRequests and FileRequests.
