+++ grizzly/users/core.py
@@ -1,0 +1,2 @@
+"""Base class for all Grizzly load user implementations."""
+
