+++ pyproject.toml
@@ -25,7 +25,6 @@ dependencies = [
     "lxml >=4.8.0,<5.0.0",
     "mypy-extensions >=0.4.3",
     "opencensus-ext-azure >=1.1.1",
-    "paramiko >=3.3.1",
     "python-dateutil >=2.8.2,<3.0.0",
     "backports.zoneinfo >=0.2.1;python_version<'3.9'",
     "pyzmq >=22.2.1,!=23.0.0",
