--- tests/unit/test_grizzly/users/base/__init__.py
@@ -1,1 +1,0 @@
-"""Commence initialization of module."""
